server/auth_domain/requirements.txt
# Core
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pydantic-settings==2.*

# Database
sqlalchemy[asyncio]==2.*
asyncpg==0.30.*
alembic==1.14.*

# Security
argon2-cffi==23.*
PyJWT[crypto]==2.*

# HTTP client (OAuth providers)
httpx==0.28.*

# Email
aiosmtplib==3.*

# Dev / Testing
pytest==8.*
pytest-asyncio==0.24.*
httpx  # also used as test client
